Fold setup into webhook init and add support for GPG
